Add example test and CircleCI config (#5)
Using Gauge for the test: https://gauge.org/

The CircleCI config could be improved for both speed and separation of
concerns (e.g. having the installation of Gauge in a separate job
from running the tests), but keeping it simple for now - no need to
optimise at this early stage.