Add e2e test using deploy preview for each commit (#16)
Any push on any branch triggers the end to end tests to run against
the specific Netlify deploy preview app [1] for that commit.

This is very powerful as it allows us to get quick feedback from
an environment which is very close to production (same https and
authentication etc).  It also means that we avoid unnecessary
complexity by, instead of trying to build the site ourselves in the
CI/CD pipeline, letting Netlify build it.

Netlify only exposes the deploy information for an individual commit in
its API [2], so this involved adding a Netlify serverless function [3]
(written in Go), to query the API to obtain the Netlify deploy url for
the commit.

This Netlify function would be useful to others, so once it has proved
its reliability then may externalise it to its own repository.

The bash scripts in the CircleCI config are in individual
files rather than multiline shell commands in `config.yml`.
This allows the build to run ShellCheck [4] against the scripts, as per
CircleCI's best practices. [5]

A future improvement may well be to migrate the scripts from Bash to Go,
which will also make testing them easier.

[1] https://www.netlify.com/blog/2016/07/20/introducing-deploy-previews-in-netlify/

[2] https://docs.netlify.com/api/get-started/

[3] https://www.netlify.com/products/functions/

[4] https://www.shellcheck.net/

[5] https://circleci.com/docs/2.0/using-shell-scripts/#shell-script-best-practices
      https://circleci.com/blog/circleci-config-teardown-how-we-write-our-circleci-config-at-circleci/