Use pre-installed Taiko
Instead of installing Taiko (and Chromium) each time, now we use
`npm link`[1][2] to link to the pre-installed global version of Taiko that is
already on the Docker image.

This speeds up the build and means there are fewer moving parts.

Was able to delete the package-lock.json as we're not actually
installing anything now during the build.

Also set the package.json to bring in any minor or patch releases
automatically (we want to test major releases properly, so good not
to accept them automatically).

[1] https://github.com/getgauge/taiko/issues/326#issuecomment-452584249
[2] https://github.com/getgauge/gauge-js/issues/174#issuecomment-435271044